wip: Move D-Bus services into separate modules
